+++ src/main/java/com/wildcodeschool/giftmefive/entity/Gift.java
@@ -5,21 +5,21 @@ import javax.persistence.JoinColumn;
 import javax.persistence.OneToOne;
 
 public class Gift {
-    private int idGift;
+    private long idGift;
     private String giftName;
     private String description;
     private float price;
     private int preference;
     private String urlImage;
     private String urlWebsite;
-    private int idList;
-    private int idFriend;
+    private long idList;
+    private long idFriend;
     @OneToOne(cascade = CascadeType.ALL)
     @JoinColumn(unique = true)
     private User user;
 
-    public Gift(int idGift, String giftName, String description, float price, int preference, String urlImage,
-                String urlWebsite, int idList, int idFriend) {
+    public Gift(long idGift, String giftName, String description, float price, int preference, String urlImage,
+                String urlWebsite, long idList, long idFriend) {
         this.idGift = idGift;
         this.giftName = giftName;
         this.description = description;
@@ -31,11 +31,11 @@ public class Gift {
         this.idFriend = idFriend;
     }
 
-    public int getIdGift() {
+    public long getIdGift() {
         return idGift;
     }
 
-    public void setIdGift(int idGift) {
+    public void setIdGift(long idGift) {
         this.idGift = idGift;
     }
 
@@ -87,19 +87,19 @@ public class Gift {
         this.urlWebsite = urlWebsite;
     }
 
-    public int getIdList() {
+    public long getIdList() {
         return idList;
     }
 
-    public void setIdList(int idList) {
+    public void setIdList(long idList) {
         this.idList = idList;
     }
 
-    public int getIdFriend() {
+    public long getIdFriend() {
         return idFriend;
     }
 
-    public void setIdFriend(int idFriend) {
+    public void setIdFriend(long idFriend) {
         this.idFriend = idFriend;
     }
 }
